server/requirements.txt
discord.py~=2.6.3
audioop-lts~=0.2.2; python_version>='3.13'
aiohttp~=3.12.15
